Move the static-pages parser logic to the fetch-as-har package
